Separando pruebas de estantes para adicionar más
